Added further designs - tower bridge, dual lambdas
Added designs with the Tower Bridge silouette, on the suggestion of Jane.

Took the concept by David, of the lambda representing a person crossing the gap
of learning.  Added a second lambda to represent either a person guiding the
student over the bridge,

Or the two lambda logos could represent a student going from beginner to a little more
experience, represented by the increasing opacity of the two lambda logos.